silx.gui.widgets.IntEdit: Add IntEdit widget (rather make it public)
