+++ doc/source/modules/gui/widgets/editwidgets.rst
@@ -1,0 +1,24 @@
+
+.. currentmodule:: silx.gui.widgets
+
+Edition widgets
+===============
+
+:mod:`FloatEdit`
+--------------------
+
+.. automodule:: silx.gui.widgets.FloatEdit
+
+.. autoclass:: silx.gui.widgets.FloatEdit.FloatEdit
+   :members:
+   :show-inheritance:
+
+
+:mod:`IntEdit`
+--------------------
+
+.. automodule:: silx.gui.widgets.IntEdit
+
+.. autoclass:: silx.gui.widgets.IntEdit.IntEdit
+   :members:
+   :show-inheritance:
